Restructure code for easier debugging
The codebase is restructured for easier debugging of the decode module.
Each method (3) are in separate files. The `finderrors.rs` generates
random data and tryies to compress and uncompress the data. Should the
file not be successfully uncompressed, the data is saved with the
specific method which failed in decompression.